toolkit/components/passwordmgr/test/mochitest/test_DOMInputPasswordAdded_fired_between_DOMContentLoaded_and_load_events.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>Test the password manager code called on DOMInputPasswordAdded runs when it occurs between DOMContentLoaded and load events</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="pwmgr_common.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <!-- In cases where the "DOMContentLoaded" event for a page has occured but not the "load" event when
  "DOMInputPasswordAdded" fires, we want to make sure that the Password Manager code (i.e.
  _fetchLoginsFromParentAndFillForm) still runs on the page.
  This scenario can happen for example when a page has very little initial HTML, but extensive JS that
  adds Custom Elements or other HTML later, or when other subresources, like images, take a while to load.
  In this test, we delay the page load with a delayed response for an image source. -->
<img src="slow_image.sjs" />
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();

// Bug 1655195: Sometimes saved logins are not removed in between mochitests when this test is run as
// part of a suite
runInParent(function removeAll() {
  let {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");
  Services.logins.removeAllLogins();
})

let readyPromise = registerRunTests();

function initLogins() {
  let {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  info("Adding one login for the page");
  let login = Cc["@mozilla.org/login-manager/loginInfo;1"].createInstance(Ci.nsILoginInfo);
  login.init("https://example.com", "https://example.com", null, "user", "omgsecret!");
  Services.logins.addLogin(login);
}

async function test_password_autofilled() {
  runInParent(initLogins);
  let numLogins = await LoginManager.countLogins("https://example.com", "https://example.com", null);
  is(numLogins, 1, "Correct number of logins");

  let formsProcessed = promiseFormsProcessed();

  info("Adding password input field to the page to trigger DOMInputPasswordAdded");
  let passwordField = document.createElement("input");
  passwordField.type = "password";
  is(document.readyState, "interactive", "Make sure 'DOMContentLoaded' has fired but not 'load'");
  document.body.append(passwordField);

  info("Waiting for the password field to be autofilled");
  await formsProcessed;

  is(passwordField.value, "omgsecret!", "Check that password field was autofilled");

  SimpleTest.finish();
}

readyPromise.then(async () => test_password_autofilled());
</script>
</body>
</html>
